requirements.txt
certifi
click  
colorama    
dataclasses    
DateTime    
Flask
Flask-Cors
importlib-metadata
itsdangerous
gunicorn
Jinja2
MarkupSafe
mysql-connector-python
pip
protobuf
pytz
setuptools
six
typing_extensions
Werkzeug
wheel
wincertstore
zipp
zope.interface
